safari-534.53-branch/LayoutTests/fast/dom/HTMLMeterElement/meter-element-form.html

<html>
<head>
<script>
if (window.layoutTestController)
    layoutTestController.dumpAsText();
</script>

<body>
<table><form id=form1>
<meter id=meter1></meter>
</form></table>
<div>This test passes if the meter element can find its form.</div>
<div id=console>FAIL</div>

<script>
if (document.getElementById('meter1').form.id == "form1")
    document.getElementById('console').innerHTML = "PASS";
</script>
</body>
</html>
