LayoutTests/fast/dom/HTMLMeterElement/meter-element-form-expected.txt

CONSOLE MESSAGE: line 16: TypeError: 'undefined' is not an object (evaluating 'document.getElementById('meter1').form.id')
This test passes if the meter element can find its form.
FAIL
